# feat: separate out write path executor with unbounded memory limit